LPS-67766 Move template resources bundles tracking to helper utility
