+++ modules/apps/foundation/portal-template/portal-template-soy/src/main/java/com/liferay/portal/template/soy/internal/SoyTemplateBundleResourceParser.java
@@ -15,26 +15,14 @@
 package com.liferay.portal.template.soy.internal;
 
 import com.liferay.portal.kernel.template.TemplateConstants;
-import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.template.TemplateResourceParser;
 import com.liferay.portal.template.URLResourceParser;
 
 import java.net.URL;
 
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
 import org.osgi.framework.Bundle;
-import org.osgi.framework.BundleContext;
-import org.osgi.framework.BundleEvent;
-import org.osgi.framework.wiring.BundleCapability;
-import org.osgi.framework.wiring.BundleWiring;
-import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Deactivate;
-import org.osgi.util.tracker.BundleTracker;
-import org.osgi.util.tracker.BundleTrackerCustomizer;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Marcellus Tavares
@@ -50,102 +38,15 @@ public class SoyTemplateBundleResourceParser extends URLResourceParser {
 	public URL getURL(String templateId) {
 		int pos = templateId.indexOf(TemplateConstants.BUNDLE_SEPARATOR);
 
-		if (pos == -1) {
-			throw new IllegalArgumentException(
-				String.format(
-					"The templateId \"%s\" does not map to a Soy template",
-					templateId));
-		}
-
-		String capabilityPrefix = templateId.substring(0, pos);
-
-		Bundle bundle = _bundleProvidersMap.get(capabilityPrefix);
-
-		if (bundle == null) {
-			throw new IllegalStateException(
-				"There are no bundles providing " + capabilityPrefix);
-		}
-
 		String templateName = templateId.substring(
 			pos + TemplateConstants.BUNDLE_SEPARATOR.length());
+
+		Bundle bundle = _soyTemplateContextHelper.getTemplateBundle(templateId);
 
 		return bundle.getResource(templateName);
 	}
 
-	@Activate
-	protected void activate(BundleContext bundleContext) {
-		int stateMask = Bundle.ACTIVE | Bundle.RESOLVED;
-
-		_bundleTracker = new BundleTracker<>(
-			bundleContext, stateMask,
-			new CapabilityBundleTrackerCustomizer("soy"));
-
-		_bundleTracker.open();
-	}
-
-	@Deactivate
-	protected void deactivate() {
-		_bundleTracker.close();
-	}
-
-	protected String getCapabilityPrefix(BundleCapability bundleCapability) {
-		Map<String, Object> attributes = bundleCapability.getAttributes();
-
-		return attributes.get("type") + StringPool.UNDERLINE +
-			attributes.get("version");
-	}
-
-	private final Map<String, Bundle> _bundleProvidersMap =
-		new ConcurrentHashMap<>();
-	private BundleTracker<List<BundleCapability>> _bundleTracker;
-
-	private class CapabilityBundleTrackerCustomizer
-		implements BundleTrackerCustomizer<List<BundleCapability>> {
-
-		public CapabilityBundleTrackerCustomizer(String namespace) {
-			_namespace = namespace;
-		}
-
-		@Override
-		public List<BundleCapability> addingBundle(
-			Bundle bundle, BundleEvent bundleEvent) {
-
-			BundleWiring bundleWiring = bundle.adapt(BundleWiring.class);
-
-			List<BundleCapability> bundleCapabilities =
-				bundleWiring.getCapabilities(_namespace);
-
-			for (BundleCapability bundleCapability : bundleCapabilities) {
-				String providerBundleKey = getCapabilityPrefix(
-					bundleCapability);
-
-				_bundleProvidersMap.put(providerBundleKey, bundle);
-			}
-
-			return bundleCapabilities;
-		}
-
-		@Override
-		public void modifiedBundle(
-			Bundle bundle, BundleEvent bundleEvent,
-			List<BundleCapability> bundleCapabilities) {
-		}
-
-		@Override
-		public void removedBundle(
-			Bundle bundle, BundleEvent bundleEvent,
-			List<BundleCapability> bundleCapabilities) {
-
-			for (BundleCapability bundleCapability : bundleCapabilities) {
-				String providerBundleKey = getCapabilityPrefix(
-					bundleCapability);
-
-				_bundleProvidersMap.remove(providerBundleKey);
-			}
-		}
-
-		private final String _namespace;
-
-	}
+	@Reference(unbind = "-")
+	private SoyTemplateContextHelper _soyTemplateContextHelper;
 
 }
